Projects no longer a collection
The projects are now written on a static page for more flexibility.
This removes the need to write long descriptions for each projects, we
can rather just link to the project website. However, it's still
possible to create dedicated pages for some projects, by just creating a
new page in the `_pages` directory (with a permalink) and linking to it.